+++ src/liberator/resource/ping/core.clj
@@ -8,22 +8,25 @@
    [liberator.mixin.hal.core :as hal-mixin]))
 
 (defn definitions
-  ([dependencies] (definitions dependencies {}))
-  ([{:keys [routes]}
-    {:keys [message]
-     :or {message "pong"}}]
-   {:handle-ok
-    (fn [{:keys [request]}]
-      (hal/add-properties
-        (hal/new-resource
-          (hype/absolute-url-for request routes :ping))
-        {:message message}))}))
+  ([{:keys [router]}]
+   {:body (fn [_] {:message "pong"})
+
+    :handle-ok
+    (fn [{:keys [request resource] :as context}]
+      (let [body-fn (:body resource)
+            body (body-fn context)]
+        (hal/add-properties
+          (hal/new-resource
+            (hype/absolute-url-for request router :ping))
+          body)))}))
 
 (defn handler
-  ([dependencies] (handler dependencies {}))
-  ([dependencies options]
+  ([dependencies]
+   (handler dependencies {}))
+  ([dependencies overrides]
    (mixin/build-resource
      (json-mixin/with-json-mixin dependencies)
      (hypermedia-mixin/with-hypermedia-mixin dependencies)
      (hal-mixin/with-hal-mixin dependencies)
-     (definitions dependencies options))))
+     (definitions dependencies)
+     overrides)))