+++ test/liberator/resource/ping/core_test.clj
@@ -1,6 +1,6 @@
 (ns liberator.resource.ping.core-test
   (:require
-   [clojure.test :refer :all]
+   [clojure.test :refer [deftest is]]
 
    [halboy.resource :as hal]
    [halboy.json :as hal-json]
@@ -11,18 +11,18 @@
 
    [liberator.resource.ping.core :as ping-resource]))
 
-(def routes
+(def router
   [""
    [["/" :discovery]
     ["/ping" :ping]]])
 
 (def dependencies
-  {:routes routes})
+  {:router router})
 
 (defn resource-handler
   ([dependencies] (resource-handler dependencies {}))
-  ([dependencies options]
-   (let [handler (ping-resource/handler dependencies options)
+  ([dependencies overrides]
+   (let [handler (ping-resource/handler dependencies overrides)
          handler (-> handler
                    ring-keyword-params/wrap-keyword-params
                    ring-params/wrap-params)]
@@ -32,32 +32,48 @@
   (let [handler (resource-handler dependencies)
         request (ring/request :get "/ping")
         result (handler request)]
-    (is (= (:status result) 200))))
-
-(deftest includes-pong-message-by-default
-  (let [handler (resource-handler dependencies)
-        request (ring/request :get "/ping")
-        result (handler request)
-        resource (hal-json/json->resource (:body result))]
-    (is (= (hal/get-property resource :message) "pong"))))
-
-(deftest includes-provided-message-when-specified
-  (let [handler (resource-handler dependencies {:message "ok"})
-        request (ring/request :get "/ping")
-        result (handler request)
-        resource (hal-json/json->resource (:body result))]
-    (is (= (hal/get-property resource :message) "ok"))))
+    (is (= 200 (:status result)))))
 
 (deftest includes-self-link
   (let [handler (resource-handler dependencies)
         request (ring/request :get "http://localhost/ping")
         result (handler request)
         resource (hal-json/json->resource (:body result))]
-    (is (= (hal/get-href resource :self) "http://localhost/ping"))))
+    (is (= "http://localhost/ping" (hal/get-href resource :self)))))
 
 (deftest includes-discovery-link
   (let [handler (resource-handler dependencies)
         request (ring/request :get "http://localhost/ping")
         result (handler request)
         resource (hal-json/json->resource (:body result))]
-    (is (= (hal/get-href resource :discovery) "http://localhost/"))))
+    (is (= "http://localhost/" (hal/get-href resource :discovery)))))
+
+(deftest includes-pong-message-in-body-by-default
+  (let [handler (resource-handler dependencies)
+        request (ring/request :get "/ping")
+        result (handler request)
+        resource (hal-json/json->resource (:body result))]
+    (is (= "pong" (hal/get-property resource :message)))))
+
+(deftest allows-body-to-be-overridden-with-data
+  (let [handler (resource-handler dependencies
+                  {:body {:message "ok"}})
+        request (ring/request :get "/ping")
+        result (handler request)
+        resource (hal-json/json->resource (:body result))]
+    (is (= "ok" (hal/get-property resource :message)))))
+
+(deftest allows-body-to-be-overridden-with-function-of-context
+  (let [handler (resource-handler dependencies
+                  {:body (fn [{:keys [request]}]
+                           (select-keys request [:request-method]))})
+        request (ring/request :get "/ping")
+        result (handler request)
+        resource (hal-json/json->resource (:body result))]
+    (is (= "get" (hal/get-property resource :requestMethod)))))
+
+(deftest allows-override-definitions
+  (let [handler (resource-handler dependencies {:exists? false})
+        request (ring/request :get "http://localhost/ping")
+        result (handler request)]
+    (is (= 404 (:status result)))))